Traslado el custom hook useSearch al directorio hooks
