+++ projects/05-react-buscador-peliculas/src/App.jsx
@@ -1,40 +1,9 @@
 import './App.css'
 import { useMovies } from './hooks/useMovies.js'
+import { useSearch } from './hooks/useSearch'
 import { Movies } from './components/Movies.jsx'
-import { useState, useEffect, useRef, useCallback } from 'react'
+import { useState, useCallback } from 'react'
 import debounce from 'just-debounce-it'
-
-function useSearch () {
-  const [search, updateSearch] = useState('')
-  const [error, setError] = useState(null)
-  const isFirstInput = useRef(true)
-
-  useEffect(() => {
-    if (isFirstInput.current) {
-      isFirstInput.current = search === ''
-      return
-    }
-
-    if (search === '') {
-      setError('No se puede buscar una película vacía')
-      return
-    }
-
-    if (search.match(/^\d+$/)) {
-      setError('No se puede buscar una película con un número')
-      return
-    }
-
-    if (search.length < 3) {
-      setError('La búsqueda debe tener al menos 3 caracteres')
-      return
-    }
-
-    setError(null)
-  }, [search])
-
-  return { search, updateSearch, error }
-}
 
 function App () {
   const [sort, setSort] = useState(false)
